+++ src/exactly_lib/default/program_modes/test_case/builtin_symbols/__init__.py
@@ -1,7 +1,5 @@
-from exactly_lib.default.program_modes.test_case.builtin_symbols import test_case_dir_symbols, strings, \
-    string_transformers
+from exactly_lib.default.program_modes.test_case.builtin_symbols import test_case_dir_symbols, strings
 
 ALL = (strings.all_strings() +
-       test_case_dir_symbols.ALL +
-       string_transformers.ALL
+       test_case_dir_symbols.ALL
        )