+++ test/exactly_lib_test/test_case_utils/string_transformers/test_resources/replace_tcds_dirs.py
@@ -4,9 +4,9 @@ from exactly_lib.test_case_file_structure import tcds_symbols
 from exactly_lib.test_case_file_structure.tcds import Tcds
 
 
-class ReplacedEnvVarsFileContentsGeneratorWithAllReplacedVariables:
+class ReplacedSymbolsFileContentsGeneratorWithAllReplacedVariables:
     """
-    Generates contents with all environment variables that this program sets.
+    Generates contents with all paths with a corresponding path symbol, and some more SDS paths.
     Some of these variables _should_ be replaced by the replacement option, and some _should_not_.
     This assures that it is tested that only the correct variables are replaced.
 
@@ -15,23 +15,23 @@ class ReplacedEnvVarsFileContentsGeneratorWithAllReplacedVariables:
     """
 
     def __init__(self):
-        self.sorted_env_var_keys = sorted(tcds_symbols.ALL_REPLACED_SYMBOLS)
+        self.sorted_symbol_keys = sorted(tcds_symbols.ALL_REPLACED_SYMBOLS)
 
     def contents_before_replacement(self, tcds: Tcds) -> str:
-        env_vars_dict = tcds_symbols.replaced(tcds)
-        values_in_determined_order = list(map(env_vars_dict.get, self.sorted_env_var_keys))
+        symbols_dict = tcds_symbols.replaced(tcds)
+        values_in_determined_order = list(map(symbols_dict.get, self.sorted_symbol_keys))
         return self._content_from_values(values_in_determined_order,
                                          tcds)
 
     def expected_contents_after_replacement(self, tcds: Tcds) -> str:
-        return self._content_from_values(self.sorted_env_var_keys,
+        return self._content_from_values(self.sorted_symbol_keys,
                                          tcds)
 
     def unexpected_contents_after_replacement(self, tcds: Tcds) -> str:
         """
         :return: Gives a variation of the expected result, that is not equal to the expected result.
         """
-        return self._content_from_values(tuple(reversed(self.sorted_env_var_keys)),
+        return self._content_from_values(tuple(reversed(self.sorted_symbol_keys)),
                                          tcds)
 
     @staticmethod
@@ -52,7 +52,7 @@ class ReplacedEnvVarsFileContentsGeneratorWithAllReplacedVariables:
                                 should_not_be_replaced_values]) + os.linesep
 
 
-class ReplacedEnvVarsFileContentsGeneratorForSubDirRelationshipBetweenHdsActAndCase:
+class ReplacedSymbolsFileContentsGeneratorForSubDirRelationshipBetweenHdsActAndCase:
     def __init__(self,
                  name_of_parent_dir__rel_hds_symbol: str,
                  name_of_sub_dir__rel_hds_symbol: str):
@@ -60,7 +60,7 @@ class ReplacedEnvVarsFileContentsGeneratorForSubDirRelationshipBetweenHdsActAndC
         self.name_of_sub_dir__rel_hds_symbol = name_of_sub_dir__rel_hds_symbol
 
     """
-    Generates contents where one of the rel-home environment variables (home/act, home/case)
+    Generates contents where one of the rel-home symbols (home/act, home/case)
     is a sub directory of the other.
     """