+++ CHANGELOG.md
@@ -25,6 +25,7 @@ with exception of a "0." prefix until version 1 is released.
  - Python compatibility: Increased to Python v 3.5.9
  - Instruction `conf`/`actor` - Remove "-" prefix from actor names
  - String transformer - Case converters - Replaces builtin symbols with primitives
+ - String transformer - TCDS dir replacement - Replaces builtin symbol with primitive
  
 ### Removed